# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: URL Diagnostics API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /monitors:
    get:
      operationId: listMonitors
      tags: [monitors]
      summary: List all monitors for the current user
      security:
        - clerkAuth: []
      parameters:
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag name(s), comma-separated
      responses:
        "200":
          description: List of monitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Monitor"
    post:
      operationId: createMonitor
      tags: [monitors]
      summary: Create a new monitor
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorInput"
      responses:
        "201":
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          description: Validation error
        "403":
          description: Plan limit reached
  /monitors/{id}:
    get:
      operationId: getMonitor
      tags: [monitors]
      summary: Get a single monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Monitor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "404":
          description: Not found
    put:
      operationId: updateMonitor
      tags: [monitors]
      summary: Update a monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorUpdate"
      responses:
        "200":
          description: Updated monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "404":
          description: Not found
    delete:
      operationId: deleteMonitor
      tags: [monitors]
      summary: Delete a monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /monitors/{id}/pause:
    post:
      operationId: pauseMonitor
      tags: [monitors]
      summary: Pause a monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
  /monitors/{id}/resume:
    post:
      operationId: resumeMonitor
      tags: [monitors]
      summary: Resume a paused monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
  /monitors/{id}/check:
    post:
      operationId: triggerMonitorCheck
      tags: [monitors]
      summary: Trigger an immediate manual check
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorCheck"
  /monitors/{id}/uptime:
    get:
      operationId: getMonitorUptime
      tags: [monitors]
      summary: Get uptime stats for a monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [24h, 7d, 30d, 90d]
            default: "30d"
      responses:
        "200":
          description: Uptime statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UptimeStats"
  /monitors/{id}/timeseries:
    get:
      operationId: getMonitorTimeseries
      tags: [monitors]
      summary: Get response time time series data
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [1h, 24h, 7d, 30d]
            default: "24h"
        - name: granularity
          in: query
          schema:
            type: string
            enum: [1m, 5m, 1h, 1d]
            default: "5m"
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeseriesData"
  /monitors/{id}/checks:
    get:
      operationId: listMonitorChecks
      tags: [checks]
      summary: List recent checks for a monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Recent checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonitorCheck"
  /monitors/{id}/incidents:
    get:
      operationId: listMonitorIncidents
      tags: [incidents]
      summary: List incidents for a specific monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Monitor incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
  /monitors/{id}/tags:
    post:
      operationId: addTagToMonitor
      tags: [tags]
      summary: Add a tag to a monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagAssignment"
      responses:
        "200":
          description: Updated monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
  /monitors/{id}/tags/{tagId}:
    delete:
      operationId: removeTagFromMonitor
      tags: [tags]
      summary: Remove a tag from a monitor
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tagId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"

  # ─── Monitor Alert Channels ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    Monitor:
      type: object
      required: [id, name, url, method, intervalSeconds, isActive, lastStatus, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        method:
          type: string
          enum: [GET, POST, HEAD]
        intervalSeconds:
          type: integer
        timeoutSeconds:
          type: integer
        expectedStatus:
          type: integer
        keywordAssert:
          type: ["string", "null"]
        regions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        notifyOnDown:
          type: boolean
        notifyOnRecovery:
          type: boolean
        lastCheckedAt:
          type: ["string", "null"]
          format: date-time
        lastStatus:
          type: ["string", "null"]
          enum: [up, down, degraded, null]
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MonitorInput:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string
          format: uri
        method:
          type: string
          enum: [GET, POST, HEAD]
          default: GET
        intervalSeconds:
          type: integer
          default: 300
        timeoutSeconds:
          type: integer
          default: 30
        expectedStatus:
          type: integer
          default: 200
        keywordAssert:
          type: string
        regions:
          type: array
          items:
            type: string
        notifyOnDown:
          type: boolean
          default: true
        notifyOnRecovery:
          type: boolean
          default: true
    MonitorUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        method:
          type: string
          enum: [GET, POST, HEAD]
        intervalSeconds:
          type: integer
        timeoutSeconds:
          type: integer
        expectedStatus:
          type: integer
        keywordAssert:
          type: string
        regions:
          type: array
          items:
            type: string
        notifyOnDown:
          type: boolean
        notifyOnRecovery:
          type: boolean

    # ─── Monitor Check ────────────────────────────────────────────────────
    MonitorCheck:
      type: object
      required: [id, monitorId, region, status, checkedAt]
      properties:
        id:
          type: string
        monitorId:
          type: string
        region:
          type: string
        status:
          type: string
          enum: [up, down, timeout, error]
        httpStatus:
          type: ["integer", "null"]
        responseTimeMs:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        checkedAt:
          type: string
          format: date-time

    # ─── Uptime ───────────────────────────────────────────────────────────
    UptimeStats:
      type: object
      required: [period, uptimePercent, totalChecks]
      properties:
        period:
          type: string
        uptimePercent:
          type: number
        totalChecks:
          type: integer
        downChecks:
          type: integer
        avgResponseTimeMs:
          type: ["number", "null"]
        p95ResponseTimeMs:
          type: ["number", "null"]
        incidentsCount:
          type: integer
        totalDowntimeMinutes:
          type: number

    # ─── Timeseries ──────────────────────────────────────────────────────
    TimeseriesData:
      type: object
      required: [monitorId, period, granularity, series]
      properties:
        monitorId:
          type: string
        period:
          type: string
        granularity:
          type: string
        series:
          type: array
          items:
            $ref: "#/components/schemas/TimeseriesPoint"
    TimeseriesPoint:
      type: object
      required: [timestamp, status]
      properties:
        timestamp:
          type: string
          format: date-time
        responseTimeMs:
          type: ["integer", "null"]
        status:
          type: string
          enum: [up, down, timeout, error]

    # ─── Incidents ────────────────────────────────────────────────────────
    Incident:
      type: object
      required: [id, monitorId, startedAt, isAcknowledged, createdAt]
      properties:
        id:
          type: string
        monitorId:
          type: string
        monitorName:
          type: ["string", "null"]
        monitorUrl:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        durationSeconds:
          type: ["integer", "null"]
        rootCause:
          type: ["string", "null"]
        affectedRegions:
          type: array
          items:
            type: string
        isAcknowledged:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Tag:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
...
    TagAssignment:
      type: object
      required: [tagId]
      properties:
        tagId:
          type: string