# lib/api-zod/src/generated/types/uptimeStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * URL Diagnostics API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UptimeStats {
  period: string;
  uptimePercent: number;
  totalChecks: number;
  downChecks?: number;
  /** @nullable */
  avgResponseTimeMs?: number | null;
  /** @nullable */
  p95ResponseTimeMs?: number | null;
  incidentsCount?: number;
  totalDowntimeMinutes?: number;
}